+++ gdsfactory/import_gds.py
@@ -1,6 +1,6 @@
 import json
 import pathlib
-from functools import cache, partial
+from functools import lru_cache, partial
 from pathlib import Path
 from typing import Optional, Tuple, Union, cast
 
@@ -281,7 +281,7 @@ def add_ports_from_labels(
 
 
 # pytype: disable=bad-return-type
-@cache
+@lru_cache(maxsize=None)
 def import_gds(
     gdspath: Union[str, Path],
     cellname: Optional[str] = None,